editautoskill and skill levels
