+++ app/src/main/java/com/mathewsachin/fategrandautomata/dagger/app/ApplicationComponent.kt
@@ -1,7 +1,6 @@
 package com.mathewsachin.fategrandautomata.dagger.app
 
 import com.mathewsachin.fategrandautomata.dagger.service.ScriptRunnerServiceComponent
-import com.mathewsachin.fategrandautomata.ui.AutoSkillItemActivity
 import com.mathewsachin.fategrandautomata.ui.MainActivity
 import com.mathewsachin.fategrandautomata.ui.auto_skill_list.AutoSkillListFragment
 import com.mathewsachin.fategrandautomata.ui.card_priority.CardPriorityActivity
@@ -18,7 +17,6 @@ interface ApplicationComponent {
 
     fun inject(into: MainActivity)
     fun inject(into: AutoSkillListFragment)
-    fun inject(into: AutoSkillItemActivity)
     fun inject(into: CardPriorityActivity)
     fun inject(into: AutoSkillItemSettingsFragment)
     fun inject(into: MainSettingsFragment)
