+++ app/src/main/res/navigation/nav_graph.xml
@@ -57,5 +57,36 @@
         android:id="@+id/autoSkillListFragment"
         android:name="com.mathewsachin.fategrandautomata.ui.auto_skill_list.AutoSkillListFragment"
         android:label="AutoSkill List"
-        tools:layout="@layout/autoskill_list" />
+        tools:layout="@layout/autoskill_list">
+        <action
+            android:id="@+id/action_autoSkillListFragment_to_autoSkillItemSettingsFragment"
+            app:destination="@id/autoSkillItemSettingsFragment"
+            app:enterAnim="@anim/fragment_open_enter"
+            app:exitAnim="@anim/fragment_open_exit"
+            app:popEnterAnim="@anim/fragment_close_enter"
+            app:popExitAnim="@anim/fragment_close_exit" />
+    </fragment>
+    <fragment
+        android:id="@+id/autoSkillItemSettingsFragment"
+        android:name="com.mathewsachin.fategrandautomata.ui.prefs.AutoSkillItemSettingsFragment"
+        android:label="Edit AutoSkill Item">
+        <argument
+            android:name="key"
+            app:argType="string" />
+        <action
+            android:id="@+id/action_autoSkillItemSettingsFragment_to_skillLevelSettingsFragment"
+            app:destination="@id/skillLevelSettingsFragment"
+            app:enterAnim="@anim/fragment_open_enter"
+            app:exitAnim="@anim/fragment_open_exit"
+            app:popEnterAnim="@anim/fragment_close_enter"
+            app:popExitAnim="@anim/fragment_close_exit" />
+    </fragment>
+    <fragment
+        android:id="@+id/skillLevelSettingsFragment"
+        android:name="com.mathewsachin.fategrandautomata.ui.prefs.SkillLevelSettingsFragment"
+        android:label="Skill Levels">
+        <argument
+            android:name="key"
+            app:argType="string" />
+    </fragment>
 </navigation>
